+++ src/ui/mod.rs
@@ -1,15 +1,7 @@
-mod big_numbers;
-mod button;
-mod checkbox;
-mod clickable_list;
-mod clickable_table;
 mod constants;
 mod gif_map;
-mod hover_text_line;
-mod hover_text_span;
-mod link_lines;
 mod panels;
-mod popup_message;
+mod renders;
 mod traits;
 mod ui_callback;
 mod ui_frame;
@@ -18,8 +10,13 @@ mod ui_screen;
 mod utils;
 mod widgets;
 
+pub(crate) use widgets::{
+    big_numbers, button, checkbox, clickable_list, clickable_table, hover_text_line,
+    hover_text_span,
+};
+
 pub use constants::UI_SCREEN_SIZE;
-pub use popup_message::*;
 pub use ui_callback::UiCallback;
 pub use ui_key::*;
 pub use ui_screen::{UiScreen, UiState};
+pub use widgets::popup_message::*;
