+++ src/ui/panels/splash_screen.rs
@@ -5,10 +5,10 @@ use crate::types::{AppResult, SystemTimeTick, Tick};
 use crate::ui::button::Button;
 use crate::ui::gif_map::*;
 use crate::ui::panels::SplitPanel;
+use crate::ui::renders::default_block;
 use crate::ui::ui_callback::UiCallback;
 use crate::ui::ui_frame::UiFrame;
 use crate::ui::utils::big_text;
-use crate::ui::widgets::default_block;
 use crate::AudioPlayerState;
 use crate::{core::world::World, store::save_game_exists};
 use core::fmt::Debug;
