+++ src/ui/widgets/hover_text_line.rs
@@ -1,4 +1,5 @@
-use super::{hover_text_span::HoverTextSpan, traits::InteractiveWidget};
+use super::hover_text_span::HoverTextSpan;
+use crate::ui::traits::InteractiveWidget;
 use itertools::Itertools;
 use ratatui::{prelude::*, widgets::Widget};
 
@@ -203,7 +204,7 @@ impl InteractiveWidget for HoverTextLine<'_> {
     fn before_rendering(
         &mut self,
         area: Rect,
-        callback_registry: &mut super::ui_callback::CallbackRegistry,
+        callback_registry: &mut crate::ui::ui_callback::CallbackRegistry,
     ) {
         let width = self.width() as u16;
         let offset = match self.alignment {
