+++ src/ui/widgets/mod.rs
@@ -1,0 +1,8 @@
+pub(crate) mod big_numbers;
+pub(crate) mod button;
+pub(crate) mod checkbox;
+pub(crate) mod clickable_list;
+pub(crate) mod clickable_table;
+pub(crate) mod hover_text_line;
+pub(crate) mod hover_text_span;
+pub(crate) mod popup_message;
