+++ src/ui/widgets/popup_message.rs
@@ -1,20 +1,19 @@
 use super::button::Button;
-use super::constants::{UiStyle, UiText};
-use super::gif_map::{self, GifMap, TREASURE_GIF};
-use super::link_lines::{render_lines_with_links, LinkAlign};
-use super::ui_callback::UiCallback;
-use super::ui_frame::UiFrame;
-use super::utils::{img_to_lines, input_from_key_event, validate_textarea_input};
-use super::widgets::{default_block, thick_block};
 use crate::core::planet::PlanetType;
 use crate::core::MAX_SKILL;
 use crate::core::{player::Player, resources::Resource, skill::Rated};
 use crate::image::utils::open_gif;
 use crate::types::*;
 use crate::ui::constants::MAX_NAME_LENGTH;
+use crate::ui::constants::{UiStyle, UiText};
 use crate::ui::gif_map::PORTAL_GIFS;
+use crate::ui::gif_map::{self, GifMap, TREASURE_GIF};
+use crate::ui::renders::{default_block, render_lines_with_links, thick_block, LinkAlign};
 use crate::ui::traits::PrintableGif;
+use crate::ui::ui_callback::UiCallback;
+use crate::ui::ui_frame::UiFrame;
 use crate::ui::ui_key;
+use crate::ui::utils::{img_to_lines, input_from_key_event, validate_textarea_input};
 use anyhow::anyhow;
 use core::fmt::Debug;
 use itertools::Itertools;
